app/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pyjwt
werkzeug
flask-cors
